Attemptfix

+++ src/me/rida/anticheat/checks/movement/FlyA.java
@@ -91,7 +91,7 @@ public class FlyA extends Check {
         if(!PlayerUtil.isOnGround(p)) {
             double distanceToGround = getDistanceToGround(p);
             double yDiff = MathUtil.getVerticalDistance(e.getFrom(), e.getTo());
-            float verbose = data.getFlyHoverVerbose();
+            int verbose = data.getFlyHoverVerbose();
 
             if(distanceToGround > 2) {
                 verbose = yDiff == 0 ? verbose + 6 : yDiff < 0.06 ? verbose + 4 : 0;
@@ -132,7 +132,7 @@ public class FlyA extends Check {
         }
         double diffY = Math.abs(from.getY() - to.getY());
         double lastDiffY = data.getLastVelocityFlyY();
-        float verboseC = data.getFlyVelocityVerbose();
+        int verboseC = data.getFlyVelocityVerbose();
         double finalDifference = Math.abs(diffY - lastDiffY);
         if(finalDifference < 0.08
                 && e.getFrom().getY() < e.getTo().getY()
